ignr(.gitignore): 更新忽略文件列表
- 添加 /results 目录到忽略列表
- 保留 /temp_batches 目录

docs: 删除 SSM 和 ASCC 详细文档

- 移除 MASA.md 文件，该文件包含 SSM 和 ASCC 机制的详细说明
- 这可能是为了简化文档结构或移除过时的信息

experiments: 添加新的联邦学习算法实验

- 在 README.md 中添加了新的实验命令，包括 Harmony 和 Cream 算法
- 更新了 CUDA_VISIBLE_DEVICES 设备编号

data: 删除 COCO 数据集子集索引文件

- 移除了 coco_subset_idx_file 二进制文件

tracking: 删除 FedMD 实验结果

- 移除了 img_txt_FedMD.csv 文件，该文件记录了 FedMD 算法的实验结果

+++ src/algorithms/MMClientTrainer.py
@@ -353,8 +353,8 @@ class MMClientTrainer(EngineBase):
             images = images.to(self.device)
             captions = captions.to(self.device)
             batch_size = images.size(0)
-            img_soft_label = torch.tensor(avg_img_logits[idx:idx+batch_size]).to(self.device)
-            txt_soft_label = torch.tensor(avg_txt_logits[idx:idx+batch_size]).to(self.device)
+            img_soft_label = torch.tensor(avg_img_logits[idx:idx+batch_size], dtype=torch.float32).to(self.device)
+            txt_soft_label = torch.tensor(avg_txt_logits[idx:idx+batch_size], dtype=torch.float32).to(self.device)
             if torch.isnan(img_soft_label).any() or torch.isnan(txt_soft_label).any():
                 print("Found nan in soft labels!")
             idx += batch_size
